Changing precision to double

+++ src/UbiProtocol.h
@@ -44,7 +44,7 @@ class UbiProtocol {
 public:
   virtual bool sendData(const char *device_label, const char *device_name,
                         char *payload) = 0;
-  virtual float get(const char *device_label, const char *variable_label) = 0;
+  virtual double get(const char *device_label, const char *variable_label) = 0;
   virtual void setDebug(bool debug) = 0;
   virtual bool serverConnected();
 };
